review code oleg & sergey

+++ refact-agent/engine/src/cloud/messages_req.rs
@@ -8,9 +8,9 @@ use tracing::warn;
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct ThreadMessage {
     pub ftm_belongs_to_ft_id: String,
-    pub ftm_alt: i32,
-    pub ftm_num: i32,
-    pub ftm_prev_alt: i32,
+    pub ftm_alt: i64,
+    pub ftm_num: i64,
+    pub ftm_prev_alt: i64,
     pub ftm_role: String,
     pub ftm_content: Option<Value>,
     pub ftm_tool_calls: Option<Value>,
@@ -25,7 +25,7 @@ pub async fn get_thread_messages(
     cmd_address_url: &str,
     api_key: &str,
     thread_id: &str,
-    alt: i64,
+    alt: i64 ,
 ) -> Result<Vec<ThreadMessage>, String> {
     let query = r#"
     query GetThreadMessagesByAlt($thread_id: String!, $alt: Int!) {
@@ -52,7 +52,7 @@ pub async fn get_thread_messages(
         "thread_id": thread_id,
         "alt": alt
     });
-    
+
     let config = GraphQLRequestConfig {
         address: cmd_address_url.to_string(),
         api_key: api_key.to_string(),
@@ -64,9 +64,9 @@ pub async fn get_thread_messages(
         config.address, thread_id, alt
     );
     execute_graphql::<Vec<ThreadMessage>, _>(
-        config, 
-        query, 
-        variables, 
+        config,
+        query,
+        variables,
         "thread_messages_list"
     )
     .await
@@ -143,7 +143,7 @@ pub async fn create_thread_messages(
         }
     }
     "#;
-    
+
     let config = GraphQLRequestConfig {
         address: cmd_address_url.to_string(),
         api_key: api_key.to_string(),
@@ -154,9 +154,9 @@ pub async fn create_thread_messages(
         config.address, thread_id, messages_len
     );
     execute_graphql_no_result(
-        config, 
-        mutation, 
-        variables, 
+        config,
+        mutation,
+        variables,
         "thread_messages_create_multiple"
     )
     .await
@@ -193,9 +193,9 @@ pub fn convert_thread_messages_to_messages(
 
 pub fn convert_messages_to_thread_messages(
     messages: Vec<ChatMessage>,
-    alt: i32,
-    prev_alt: i32,
-    start_num: i32,
+    alt: i64,
+    prev_alt: i64,
+    start_num: i64,
     thread_id: &str,
     user_preferences: Option<Value>,
 ) -> Result<Vec<ThreadMessage>, String> {
@@ -205,7 +205,7 @@ pub fn convert_messages_to_thread_messages(
         delay_images.clear();
     };
     for (i, msg) in messages.into_iter().enumerate() {
-        let num = start_num + i as i32;
+        let num = start_num + i as i64;
         let mut delay_images = vec![];
         let mut messages = if msg.role == "tool" {
             let mut results = vec![];
